+++ octopus/exchange/materialize/jsonbackend/show.go
@@ -10,12 +10,12 @@ import (
 
 // ShowWorker struct handle encode
 type ShowWorker struct {
-	Supplier  string `json:"supplier"`
-	Publisher string `json:"publisher"`
-	Demand    string `json:"demand"`
-	Winner    int64  `json:"winner"`
-	Time      string `json:"time"`
-	Profit    int64  `json:"profit"`
+	Supplier  string  `json:"supplier"`
+	Publisher string  `json:"publisher"`
+	Demand    string  `json:"demand"`
+	Winner    float64 `json:"winner"`
+	Time      string  `json:"time"`
+	Profit    float64 `json:"profit"`
 }
 
 type show struct {
@@ -63,14 +63,14 @@ func (*show) Report() func(error) {
 }
 
 // ShowJob return a broker job
-func ShowJob(demand string, IP string, winner int64, t string, supplier string, publisher string, profit int64) broker.Job {
+func ShowJob(demand string, IP string, winner float64, t string, supplier string, publisher string, profit float64) broker.Job {
 	return &show{
 		inner: fillShowJob(demand, winner, t, supplier, publisher, profit),
 		key:   IP,
 	}
 }
 
-func fillShowJob(demand string, winner int64, t string, supplier string, publisher string, profit int64) *ShowWorker {
+func fillShowJob(demand string, winner float64, t string, supplier string, publisher string, profit float64) *ShowWorker {
 	return &ShowWorker{
 		Publisher: publisher,
 		Time:      t,
